+++ packages/server/src/discord/components/presets/index.ts
@@ -1,6 +1,7 @@
 import { AnnouncementComponentPresets } from "./announcements";
 import { CommonComponentPresets } from "./common";
 import { GalleryComponentPresets } from "./gallery";
+import { HallOfFameComponentPresets } from "./hall-of-fame";
 import { ModpackChangelogComponentPresets } from "./modpack-changelog";
 
 /** Aggregated Components V2 presets: common presets are spread at the top level, domain presets are nested */
@@ -8,5 +9,6 @@ export const ComponentPresets = {
   ...CommonComponentPresets,
   announcements: AnnouncementComponentPresets,
   gallery: GalleryComponentPresets,
+  hallOfFame: HallOfFameComponentPresets,
   modpackChangelog: ModpackChangelogComponentPresets,
 };